Remove all lock handling from KSharedDataCache::SharedMemory.
KSDC::SharedMemory corresponds to the part of the cache that is shared between
processes. However locking and unlocking is always done in the context of a
particular process (under the control of KSharedDataCache::Private) so move the
lock-handling there, as this allows a process to check once for the type of
locking used on a cache and remember that, whereas virtuals or function
pointers are not supported in shared memory.

svn path=/trunk/KDE/kdelibs/; revision=1181975