Optimize locking a shared cache.
KSharedDataCache uses a process-local mutex to protect against multiple threads trying
to remap an improperly sized cache. But this pretty much never happens, and the locking
can be better done within the code that does the remapping. This refactoring removes the
cost of a QMutexLocker creation and lock acquisition from the common lock path.

svn path=/trunk/KDE/kdelibs/; revision=1183338