Fix warnings "QPainter::begin: Widget painting can only begin as a result of a paintEvent" on pages with form widgets (testcase: https://bugs.kde.org/query.cgi).
The reason is that viewportEvent() is supposed to draw on the viewport,
but khtmlview would try to call paintEvent() to draw on widget() from there,
and QPainter didn't expect that, so nothing gets drawn anyway.
Clearly the painting is happening anyway, from the eventFilter on widget(),
so this isn't needed. Confirmed by Germain Garand long ago.

svn path=/trunk/KDE/kdelibs/; revision=1185322