Backport a "valid page" check for KSharedDataCache to KDE 4.5.3.
This adds additional checks that seem to prevent crashes when changing the system time.
This only masks the symptom though, it should be a logic error to be removing a page
that is already marked as free in the code paths that could reach the affected function.
Either way, it's one less crash.

I'll leave the bug open until I feel I've figured out why the underlying cause of the
bug when changing system time.

As an aside to Beat Wolf, I am just as likely (if not more) to lose patches on
ReviewBoard as I am on bugs.kde.org.  Adding BugZilla attachments to
ReviewBoard isn't necessary for that reason, ReviewBoard should be for things
that require actual code review. ;)

CCBUG:253795

svn path=/branches/KDE/4.5/kdelibs/; revision=1187999