KCalendarSystem: Add convenience functions for first/last day of year/month
Add various convenience functions to return the first or last day of the month
or year.  It is possible to work these out using the main api, but it's rather
long-winded and ugly and easy to get wrong.

Developed in response to a need in Nepomuk, see  review on 
http://svn.reviewboard.kde.org/r/5671/

CCMAIL: trueg@kde.org


svn path=/trunk/KDE/kdelibs/; revision=1188685