KLocale: Add Country Subdivision Code setting
Add a setting to KLocale for the ISO 3166-2 Country level 2 code, i.e. country
divisions like States and Provinces.

For 4.6 this is simply a dumb repository for the value with no validation or
translation services provided.  Code using the iso-codes project as a source
for validation and translation is being developed in kdepimlibs and kdebase
and will be moved into kdelibs in 4.7 once the api requirements and dependency
implications are better understood.


svn path=/trunk/KDE/kdelibs/; revision=1189361