Add timestamping to KSharedDataCache to support KGameRenderer and similar libraries.
As per the KDE Platform 4.6 feature plan, KSharedDataCache now supports
timestamping.  Timestamping is kind of a misnomer as it's really just a "do
whatever" shared int that is not used by KSharedDataCache at all (although
KSharedDataCache will pre-fill it with ::time(0)). However this restores a
feature that was in KPixmapCache and used in kdegames.

This has been tested with a simple test application to verify that both setting and
retrieving the timestamp work. Sample code is included in the API documentation. All the
tests that I could get to pass before starting still pass (and apparently I need to work
on getting a 'straight board' before further kdelibs development ;)

svn path=/trunk/KDE/kdelibs/; revision=1190832