Add KLocalizedDate to make localization of dates fun.
The new KlocalizedDate class is a convenience wrapper around KCalendarSystem
and QDate to make it easy for coders to localize dates.  The api is very
consistent with QDate making most uses a simple changing of class name.

See review http://svn.reviewboard.kde.org/r/5692/

In making the new api as simple and easy to use, a number of new features were
added to KLocale and KCalendarSystem, mostly convenience enums, but also Week
Numbering system (currently only ISO Week supported).

The apidox is not yet complete, but I will finish this after hard freeze kicks
in.


svn path=/trunk/KDE/kdelibs/; revision=1193650