Be more correct with a return type in KSharedDataCache.
Vishal Rao noticed that a lock validity check returned false in a function
expected to return unsigned. (It actually happened twice :( )

C++ converts the value false to the proper value (0), which happened to be
correct as an unsigned int here as well (it even compiled with no warning). I'm
correcting this for the sake of posterity however, although there should be no
user-visible effect.

This is in the 4.5 branch (4.5.5 release), I will verify this isn't also a
problem in trunk. Thanks to Vishal Rao for noticing.

svn path=/branches/KDE/4.5/kdelibs/; revision=1203424