Store a pointer to the KLocale config file for later reference.
Currently KLocale does not save a pointer to the config file that was used to
initialise it.  However, because the creation of the KCalendarSystem is delayed
until it is actually used the config file is not available when the calendar
needs it to read the per-calendar system user settings.  Hence the need to keep
a copy of the pointer for later reference.  In the vast majority of cases the
pointer will be null indicating that the KGlobal::config() should be used.  The
only cases in KDE svn where a config is passed in are the Locale KCM and Kraft.

See reviewboard http://svn.reviewboard.kde.org/r/6047/


svn path=/trunk/KDE/kdelibs/; revision=1204354