Remove kDebug()-type methods from KSharedDataCache::Private::mapSharedMemory()
This change should make me more popular, as it prevents crashes from using kDebug before
a global component has been set. And it's probably for the best -- this code is called
very early in a kdelibs-user's startup it seems, before all the debug areas and such are
ready for use anyways.

In case there really are problems, we just fallback to Qt debugging facilities here.

svn path=/trunk/KDE/kdelibs/; revision=1204383