Fix KDateTimeFormatter to not create an en_US locale unless really needed.
Creating the en_US locale is very inefficent as it triggers an updateCatalogs()
every time, so creating the locale every time a date is parsed is horrifically
inefficient.  Delaying the creation until it is actually needed and keeping it
around afterwards is a temporary solution to fix the most common use case, but
is not a long term solution when KDateTime starts using it.


svn path=/trunk/KDE/kdelibs/; revision=1206679