an entire pixmap cache just to cache the values of the group sizes! it used to be used for many more things, which made this more sensible. not so much anymore.
what this does mean is that ~37ms of CPU time (on my laptop anyways) is
moved up into KIconLoader's init rather than on first icon fetch. what is saved
is the time to create the icon cache itself (negligable: ~2ms on my system)
and the memory overhead.

this "frees" KIconLoader from KIconCache and therefore also KPixmapCache.
we can now remove the class in "KDE5" as noted in the comments in the header.

svn path=/trunk/KDE/kdelibs/; revision=1213567