Disabling the UPnP backend in 4.6. :-/
It works fine on libsolid side, but for real world user the whole
experience sucks because: a) the upnp ioslave is not in kdebase and b)
usability wise they get flooded by too many devices higher up in the
stack.

Work is not lost, we'll try hard to get it right in 4.7.

svn path=/branches/KDE/4.6/kdelibs/; revision=1214021