Merged revision:r1215528 | orlovich | 2011-01-18 18:46:40 -0500 (Tue, 18 Jan 2011) | 11 lines
Fix problems with negative indeces in bracket ops in cases we can statically
determine arguments to be of int32 type. Fixes reply problem on 
userbase.kde.org. I think I'll probably just have to bite the bullet and add in 
a uint32 type to FrostByte's type system, but getting conversion rules right
for that will be tricky... and it'll be extra-annoying that all the bitops 
like to produce int32.

BUG:262398
BUG:204143

svn path=/branches/KDE/4.6/kdelibs/; revision=1215529