Merged revision:r1215544 | orlovich | 2011-01-18 19:59:02 -0500 (Tue, 18 Jan 2011) | 18 lines
Various fixes for the new twitter website:
1) The alleged DOM 0 DOMNode::item doesn't exist in Gecko and WebCore at least,
and it confuses the script loader (as we have NodeList-like .item, but not an 
index [[Get]]). Remove it, but add one to HTMLSelectElement as that's legit 
HTML5.

2) When document.domain is set to the present domain, we still want to call
setDomainFromDOM(), as it would set the 'script-set domain' bit moving the
origin into that namespace --- otherwise it will be inaccessible to other
things that opted in.

3) XMLHttpRequest should complete relative URLs in its own context, not the
dynamic one.

BUG: 261041

svn path=/branches/KDE/4.6/kdelibs/; revision=1215545