Make the KProcess mode for KIO also usable on OSX and use it.
On OSX, as of 10.6, one cannot fork with out exec'ing, more or less,
which makes the way kdeinit works problematic. In order not to introduce
yet another code path, this makes the so far Windows specific way of
handling kio processes via KProcess, instead of kdeinit available
more generically and uses it on OSX as well. This is based on
proof of concept patches by David Faure.

CCBUG: 209903