KListWidgetSearchLine: filter items that are inserted or changed
Up to now, the search line only filters items that are in the widget at
the time a search is started. With this commit, items that are added
or modified later are filtered as well. Unit test included

Cherry-picked from commit abc9c4c67cf5a25d0f279beedc81727cca154cad

BUG: 265709
FIXED-IN: 4.6.1