Toolbar-buttons: Allow easier configuration of text-visibility
When "Text Alongside Icons" is selected it is possible already to configure for each toolbar-button whether the text should be shown. However the configuration setting is quite hidden under Configure Toolbars... -> Change Text... -> [x] "Hide text when toolbar shows text alongside icons" and some users where not aware about such a setting at all (see bug 257918 and duplicates).

The patch provides an additional entry in the context-menu that allows to switch the text-visibility of a button by one click.

CCBUG: 257918