Add paths to qt plugins (from kstandarddirs) before qapp is created.
Necessary for apps started through d-bus activation, so that the kde widget
style is loaded. The QCoreApp::instance test was "so that qt.conf is read" (db852684)
but the whole idea of 9d7eb8f7 is that this doesn't rely on the qt config files anymore.

FIXED-IN: 4.6.2
BUG: 267770
(cherry picked from commit 979a19afe93d1e4a5be684c1139b6ab242e6b9b6)