KNewFileMenu: Add clear buttons to dialogs asking for a name
This commit adds a clear button to the KLineEdits in the dialogs
that ask the user for a file or folder name when a file or folder
is to be created. Note that the KLineEdit text is set after the
clear button is enabled. If it is done the other way around, the
clear button is not shown initially.

CCBUG: 267581
FIXED-IN: 4.6.2
(cherry picked from commit 85878a6acfb5f03ec06f58340296582ba9aab7e3)