KNewFileMenu: Add clear buttons to dialogs asking for a name
This commit adds a clear button to the KLineEdits in the dialogs
that ask the user for a file or folder name when a file or folder
is to be created. Note that the KLineEdit text is set after the
clear button is enabled. If it is done the other way around, the
clear button is not shown initially.

BUG: 267581
FIXED-IN: 4.6.2