Emit the finished signals after calling ftpCloseCommand.
This fixes the problem where clicking on a file while browsing an FTP site does
not show its content. Instead only the associated application is launched and
either an error message, in case of kate, and/or a notification about the completion
of file copying is shown.

CCBUG:203445
CCBUG:181393
FIXED-IN:4.6.2

  (cherry picked from commit a1597d87a46e951fd62b632f250ee7ef223b51e4)