Show the api.kde.org's search box with --api-searchbox.
If --api-searchbox (which is off by default) is passed to doxygen.sh,
each apidox page will have api.kde.org's search box below the rest of
the left menu.

There is no way not to hardcode the module list and not duplicate the
searchbox found in /trunk/quality/website/api.kde.org for now.

Reviewed-by: Allen Winter
BUG: 254496