Move the api.kde.org search box HTML to a separate file.
Hardcoding that much of HTML code in a shell script was very ugly, and
was even more so due to the amount of escaping that needed to be done
(and there was no way we could use single quotes inside the string, so
the generated inline JavaScript code was also wrong).

The searchbox code now lives in api_searchbox.html, which means it is
easier for projects to customize them as well.

CCMAIL: winter@kde.org