Increase intuitiveness of KFileDialog filtering.
This patch attempts to improve the usefulness and usability of
KFileDialog's filter field when in Opening mode.

Firstly, if the filter text isn't:
 * the display name of one of the filters passed to the dialog or
 * one or more space separated mimetype specifiers (containing a '/') or
 * one or more space separated file globs (containing '*', '?' or '['),
we convert the text to a glob by prepending and appending asterisks.
This lets the user enter a piece of text (without having to know any
glob patterns) and see only the files whose names contain that text,
much the same as they would when filtering in Dolphin.

Secondly, the filtering updates on the fly as the filter text is typed.
Previously, the filtering updated only when Return was pressed, which
differs from the behaviour of most of KDE's other filter boxes. The old
behaviour is especially confusing when one clicks the small clear button
embedded in the combobox, because it clears the box, but the filtering
is unchanged until the user goes to the keyboard to press Return.

REVIEW:101013
FEATURE:142900
FIXED-IN:4.7
DIGEST: