Proper "All files" option when using mime types as filters
Patch by Thomas Fischer.
If the developer adds a mimetype 'all/allfiles' to his/her list of mime
types for filtering in e.g. KFileDialog::getOpenUrl's second parameter,
setMimeFilter recognizes the request for an "all files" filter option
and adds it to the list of options in the combobox.
REVIEW: 101075