KCalendarSystem - Narrow Month/Day Name Support
Add support for Narrow Name format for Month and Week Day Names. This
can be obtained either using the standard monthName() and weekDayName()
functions with the new KCalendarSystem enum values, or using the
formatDate variant with the KLocale::NarrowName enum.

Note that the Narrow Name is not unique and so can only be used for
display and not input.

Clean up the month and weekday i18n calls to use a standard context
format, and tweak a few of the names used.  Hopefully not too much
work for the translators.

FEATURE:
CCMAIL: iamsergio@gmail.com