Fix Date Parser to use correct year for month name
Fixes bug that doesn't get triggered currently, but new changes coming
does.  Was calculating year to use if year not yet parsed, just wasn't
actually using it.  Up to now that was fine as the monthName() function
didn't care, but next change does a validity check and it was failing.