Add declarative image provider
Simple image provider that will provide KDE Icons when we use
the url: "image://icon/$ICONAME" as the source of an Image in QML.

This allow us to load any KDE icon. We still need to support modes/states
of icons, but that will come tomorrow's morning :P Should be really fast
to implement tough.

Signed-off-by: Artur Duque de Souza <asouza@kde.org>