A bit more sanity checking for loading a cache.
I've been running this for about a month now with no ill effects, which
should make sense given that it's just making it more restrictive to
load a cache from disk. There's no specific bug that would be fixed with
this, although I started investigating this based on an issue Ian Wadham
had been having on his system.