Provide a confirmation dialog for scripted remote widgets
Make really really sure the user wants to install the widget on his
system. Indeed, they're scripted and the user might not trust the source
so provide some information on them and the ability to cancel the
action.

Implementation wise it uses a KMessageBox from a job which is really not
great but the best we can do so far (remote widget support needs a
refactoring to be able to improve that). Also, the bad bad man forced
me to use a HTML table in the message box label to align properly the
information... yes, aseigo, looking at you... the 90s are calling their
HTML tables back!