Revert "A bit more sanity checking for loading a cache."
This reverts commit cdacd60876d78ffb4dee0990ea9f01382e24b9b8.

Maksim Orlovich noted that there was much /less/ sanity checking in the
"new initial cache" case. I'm not sure how I missed that given I was
testing for so long I only accidentally came across the patch a month
later via a git-status. Either way the safest course of action is to
revert.