Solid: More efficient SSE3/SSE4 checks.
This moves the SSE3 and SSE4 checks from the udev backend to the shared code,
reusing the existing cpuid testing framework. It also avoids the overhead of
parsing /proc/cpuinfo, which was being done every single time
Processor::instructionSets was called (because the code was not part of the
initialization of the cached static variable).

I changed the flag internally used for Altivec from 0x1 to 0x2 to avoid overlap
with the new x86 flags.

Discussed on IRC with Lukáš Tinkl.