KLocale: Fix readMoney() and readNumber() to handle positive signs
Both formatMoney() and formatNumber() assume that the positive sign is a
null and so if a positive sign is set they don't remove it during the
parse. This is ok in formatNumber if the symbol is + as the Qt toDouble
call will handle it, but any other symbol in any other position as is
possible in formatMoney will fail.

Add unit tests for formatMoney and readMoney as there were none, this
is how I found the bug.  Hard to belive in over 11 years since the code
went in that no-one bothered writing the tests or spotted the problem.