move the optionality of experimental libs into the experimental/CMakeLists.txt file
this is due to internal dependencies being created (e.g. libplasma to libkdeclarative)
that require some of the experimental lib additions to be built for kdelibs internal
use, even though those experimental libs are still not part of the public API.

thanks to Timo Gurr for noting the problem and helping provide a fix.

BUG:272280