kdeui: New Date/Time widgets
The KDateEdit and KTimeEdit widgets from KDEPIM have proven very
popular with copies being made by at least 6 other apps.  It has long
been planned to move these widgets into kdelibs to save duplication
and finally they are here.

Well, almost.  These are clean re-writes from scratch designed to
replicate the behaviour of the old widgets but with cleaner code and
api.  As such they are still a little rough and it is recommended
not to switch to these until beta 2 when a full review will have been
completed.

KDateComboBox:  A combo box for entering a single date, with a drop-
                down date picker.

KTimeComboBox:  A combo box for entering a single time, with a drop-
                down list of times.

KDateTimeEdit:  A KDateTime based widget for choosing a timezone aware
                date, it embeds the date and time combos but also adds
                optional time zone and calendar system combos.

KDateEdit:      A convenience subclass of KDateTimeEdit for date only

KTimeEdit:      A convenience subclass of KDateTimeEdit for time only

The time zone and calendar combos will later be split out for separate
use.

All widgets are available in the KDE Qt Designer plug-in.

And yes, just made the freeze, yet again!