Revert wrong changes from 22f18f19cc8234456efd94f99fd7c4538de6010c.
...and fix the unused variable warning the right way.
The extra variable is necessary because the value it is copied from
changes, which is the whole point of the assertions.

Rename HostQueue::nextStartingJob() to takeFirstInQueueue() which
(correctly) suggests that it mutates its object. I was never quite
happy with the previous name, which contributed to the mistake fixed
here.

CCMAIL: adawit@kde.org