Add a means of retrieving the current index while creating widgets.
KWidgetItemDelegate was designed with the idea that all items in a
model would want the same type of widget, but that isn't desired in a
QTreeView where some items are parents that span an entire row. API
design limitations prevent adding an argument, so instead a QModelIndex
property named "goya:creatingWidgetForIndex" is set on the delegate
before createItemWidgets is called.

REVIEW: 101219