Cache values for offsetAtZoneTime() and offsetAtUtc()
As pointed out by Sérgio Martins, callgrind shows that
KTimeZone::offsetAtZoneTime() uses a lot of CPU. The time zone
transition is now cached so that when it is called for successive
dates in the same year and season, the processing is minimal.

Cache use has been tested and the unit test expanded.