Fix crash in KDirOperator::Private::_k_slotExpandToUrl(...)
When entering '/' in the 'Name' line of the file open dialog, pressing
backspace and then pasting 'a/b', where a is a subfolder of the current
folder and b a file in that folder, KDirOperator may crash because it
accesses a null KFileItem. This commit fixes the crash. Unit test
included.
CCBUG: 187066