Fix crash in KDirOperator::Private::_k_slotExpandToUrl(...)
When entering '/' in the 'Name' line of the file open dialog, pressing
backspace and then pasting 'a/b', where a is a subfolder of the current
folder and b a file in that folder, KDirOperator may crash because it
accesses a null KFileItem. This commit fixes the crash. Unit test
included.
BUG: 187066
FIXED-IN: 4.6.4
(cherry picked from commit 8b5c38c8d9fae85e002fae9f325277b5b200d44d)