Fix DBus timeout when kbuilsycoca takes too much time to run
... or when a kuniqueapplication takes time to start.

This should fix a large number of bugs where ksycoca code crashes
due to not waiting enough for kbuildsycoca.

I'll also get a patch into future versions of Qt to make the timeout
configurable in QDBusAbstractInterface, so that we don't have to port
all the possibly-long dbus calls to QDBusConnection like in this patch.

CCBUG: 211401