kservicetest: Wait for the KSycoca databaseChanged DBus signal.
The test was adding a service file and running kbuildsycoca4 with KDEHOME
set, but the process doesn't re-read sycoca until the databaseChanged
signal is received, and it won't be received if the event loop doesn't
spin a bit. So the new sycoca with the new service file wasn't being
loaded.

In addition, it's a nice addition to the test to check whether the signal
actually gets sent :)
(cherry picked from commit 8407f0fc845fe724d53c17a2d5b2c09743411a9e)