local.l10n.xml: load the correct file, adapt the configuration
Previously, local.l10n.xml was defined to use a build-time generated
file containing the a list of languages for which we provides
customizations (see kdoctools/customization/xsl). It seems that this
is not the intended way for customization to work; it used to work by
accident for docbook-xsl<=1.75 but the new optimized l10n structure
introduced with docbook-xsl>=1.76 broke it (see:
http://lists.oasis-open.org/archives/docbook-apps/201009/msg00143.html )

The idea now is to directly load the correct customization file for
the required language.

WARNING: in order to make this change working (especially the
changes to kdoctools/customization/kde-include-common.xsl.cmake)
few changes have been performed to the files under
kdoctools/customization/xsl. For each of them:
- the <?xml ... ?> header was removed
- the opening tag like:
  <l:l10n xmlns:l="http://docbook.sourceforge.net/xmlns/l10n/1.0" language="en" english-language-name="English">
  was changed as follow:
  <l:i18n xmlns:l="http://docbook.sourceforge.net/xmlns/l10n/1.0">
  <l:l10n language="en" english-language-name="English">
- the corresponding closing tag for <l:i18n> was added at the end of
  the file.

REVIEW:101791
CCBUG: 271331
CCMAIL: kde-docbook@kde.org
CCMAIL: kde-i18n-doc@kde.org