Fix null/empty confusion in KUrl::htmlRef() and its unittests.
Patch by Andrea Iacovitti <aiacovitti@libero.it>, who wrote:

This aims to be a fix for bug #276204 related to htmlRef(), and the kurl unit
test that return a false positive because of QCOMPARE treating QString("") == QString()
After fixing the unit test related to url reference stuff, 2 of them failed.
One related to htmlRef() (as in #276204),
the other related to setRef() not able to set Empty (but not Null) ref.
-> fixed setRef() too.
CCBUG: 276204