Export the function isShiftAsModifierAllowed()
This function inside KKeySequnceWidgetPrivate is needed by the
kglobalaccel daemon, export here rather than duplicating code and risk
falling out of sync.
Also add Qt::Key_Backtab to the list, this is not needed internally
since there is a special case for Alt+Shift+Tab to not be recorded as
Alt+Backtab but other users of this function will need it.
CCMAIL: kde@michael-jansen.biz
REVIEW: 101515