Use platform palette and fonts
If we are not running a full session, let the platform plugin set the
palette and fonts. This means KDE palette and fonts are used when a KDE
application is running on a KDE desktop, but the GNOME palette and fonts
are used when running on a GNOME desktop.

Based on a patch suggested by Olivier Goffart.