kshareddatacache: Remove unused-but-set variable.
adawit noticed that this variable trips a warning in gcc since
freedPagesRequired was set but never actually used later.

In this case it's not a bug, as the intention was to be an optimization
to avoid freeing too many cached entries by instead freeing the minimum
amount necessary. However, the entry-freeing loop already works that
way, so this code was not required.