KFilePreviewGenerator: update JPG preview plugin settings if needed
In KDE <= 4.6, the 'jpegrotatedthumbnail' plugin was used to provide
thumbnails for JPG images which are rotated correctly. In KDE 4.7, this
is done by the usual 'jpegthumbnail' plugin if it is configured
accordingly. This patch adds a check to KFilePreviewGenerator's
constructor that replaces the obsolete plugin by the correct one.

CCBUG:277251
(cherry picked from commit e3c0ddfa5b4c0085687884a4e385241b2d86f7c4)