KFilePreviewGenerator: update JPG preview plugin settings if needed
In KDE <= 4.6, the 'jpegrotatedthumbnail' plugin was used to provide
thumbnails for JPG images which are rotated correctly. In KDE 4.7, this
is done by the usual 'jpegthumbnail' plugin if it is configured
accordingly. This patch adds a check to KFilePreviewGenerator's
constructor that replaces the obsolete plugin by the correct one.

BUG:277251
FIXED-IN:4.7.0