Improve stability of kparts/tests/openorsavequestion_unittest
The test makes some assumptions on apps that are associated with certain
mime types. If these assumptions are wrong for the system the test is
run on, the test fails and then crashes, see, e.g.,
http://my.cdash.org/testDetails.php?test=6683060&build=208738
This commit skips the affected tests in that case and outputs a warning.
(cherry picked from commit 62998d94d50a58ce4398725b73af6e5ea9ea5076)